+++ esphome/components/modbus_controller/modbus_controller.h
@@ -219,7 +219,7 @@ template<typename N> N mask_and_shift_by_rightbit(N data, uint32_t mask) {
   if (result == 0) {
     return result;
   }
-  for (int pos = 0; pos < sizeof(N) << 3; pos++) {
+  for (size_t pos = 0; pos < sizeof(N) << 3; pos++) {
     if ((mask & (1 << pos)) != 0)
       return result >> pos;
   }
