+++ esphome/components/st7735/st7735.cpp
@@ -446,7 +446,7 @@ void HOT ST7735::write_display_data_() {
   this->dc_pin_->digital_write(true);
 
   if (this->eightbitcolor_) {
-    for (int line = 0; line < this->get_buffer_length(); line = line + this->get_width_internal()) {
+    for (size_t line = 0; line < this->get_buffer_length(); line = line + this->get_width_internal()) {
       for (int index = 0; index < this->get_width_internal(); ++index) {
         auto color332 = display::ColorUtil::to_color(this->buffer_[index + line], display::ColorOrder::COLOR_ORDER_RGB,
                                                      display::ColorBitness::COLOR_BITNESS_332, true);
